create user.controller/helpers.js
